Converts ManagedServerRepository into Account
