+++ src/server_manager/model/account.ts
@@ -1,0 +1,34 @@
+// Copyright 2020 The Outline Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import {ManagedServer, RegionId} from "./server";
+
+// Keys are cityIds like "nyc".  Values are regions like ["nyc1", "nyc3"].
+export type RegionMap = {
+  [cityId: string]: RegionId[]
+};
+
+// Repository of ManagedServer objects.  These servers are created by the server
+// manager on cloud providers where we can provide a "magical" user experience,
+// e.g. DigitalOcean.
+export interface Account {
+  // Lists all existing Shadowboxes. If `fetchFromHost` is true, performs a network request to
+  // retrieve the servers; otherwise resolves with a cached server list.
+  listServers(fetchFromHost?: boolean): Promise<ManagedServer[]>;
+  // Return a map of regions that are available and support our target machine size.
+  getRegionMap(): Promise<Readonly<RegionMap>>;
+  // Creates a server and returning it when it becomes active (i.e. the server has
+  // created, not necessarily once shadowbox installation has finished).
+  createServer(region: RegionId, name: string): Promise<ManagedServer>;
+}
