+++ src/server_manager/web_app/digitalocean_server.ts
@@ -20,17 +20,12 @@ import * as errors from '../infrastructure/errors';
 import {asciiToHex, hexToString} from '../infrastructure/hex_encoding';
 import * as do_install_script from '../install_scripts/do_install_script';
 import * as server from '../model/server';
+import {Account} from '../model/account';
 
 import {ShadowboxServer} from './shadowbox_server';
 
-// WARNING: these strings must be lowercase due to a DigitalOcean case
-// sensitivity bug.
-
-// Tag used to mark Shadowbox Droplets.
-const SHADOWBOX_TAG = 'shadowbox';
 // Prefix used in key-value tags.
 const KEY_VALUE_TAG = 'kv';
-
 // The tag key for the manager API certificate fingerprint.
 const CERTIFICATE_FINGERPRINT_TAG = 'certsha256';
 // The tag key for the manager API URL.
@@ -64,7 +59,7 @@ enum InstallState {
   DELETED
 }
 
-class DigitaloceanServer extends ShadowboxServer implements server.ManagedServer {
+export class DigitalOceanServer extends ShadowboxServer implements server.ManagedServer {
   private eventQueue = new EventEmitter();
   private installState: InstallState = InstallState.UNKNOWN;
 
@@ -326,108 +321,4 @@ function startsWithCaseInsensitive(text: string, prefix: string) {
 
 export function GetCityId(slug: server.RegionId): string {
   return slug.substr(0, 3).toLowerCase();
-}
-
-const MACHINE_SIZE = 's-1vcpu-1gb';
-
-export class DigitaloceanServerRepository implements server.ManagedServerRepository {
-  private servers: DigitaloceanServer[] = [];
-
-  constructor(
-      private digitalOcean: DigitalOceanSession, private image: string, private metricsUrl: string,
-      private sentryApiUrl: string|undefined, private debugMode: boolean) {}
-
-  // Return a map of regions that are available and support our target machine size.
-  getRegionMap(): Promise<Readonly<server.RegionMap>> {
-    return this.digitalOcean.getRegionInfo().then((regions) => {
-      const ret: server.RegionMap = {};
-      regions.forEach((region) => {
-        const cityId = GetCityId(region.slug);
-        if (!(cityId in ret)) {
-          ret[cityId] = [];
-        }
-        if (region.available && region.sizes.indexOf(MACHINE_SIZE) !== -1) {
-          ret[cityId].push(region.slug);
-        }
-      });
-      return ret;
-    });
-  }
-
-  // Creates a server and returning it when it becomes active.
-  createServer(region: server.RegionId, name: string): Promise<server.ManagedServer> {
-    console.time('activeServer');
-    console.time('servingServer');
-    const onceKeyPair = crypto.generateKeyPair();
-    const watchtowerRefreshSeconds = this.image ? 30 : undefined;
-    const installCommand = getInstallScript(
-        this.digitalOcean.accessToken, name, this.image, watchtowerRefreshSeconds, this.metricsUrl,
-        this.sentryApiUrl);
-
-    const dropletSpec = {
-      installCommand,
-      size: MACHINE_SIZE,
-      image: 'docker-18-04',
-      tags: [SHADOWBOX_TAG],
-    };
-    return onceKeyPair
-        .then((keyPair) => {
-          if (this.debugMode) {
-            // Strip carriage returns, which produce weird blank lines when pasted into a terminal.
-            console.debug(
-                `private key for SSH access to new droplet:\n${
-                    keyPair.private.replace(/\r/g, '')}\n\n` +
-                'Use "ssh -i keyfile root@[ip_address]" to connect to the machine');
-          }
-          return this.digitalOcean.createDroplet(name, region, keyPair.public, dropletSpec);
-        })
-        .then((response) => {
-          return this.createDigitalOceanServer(this.digitalOcean, response.droplet);
-        });
-  }
-
-  listServers(fetchFromHost = true): Promise<server.ManagedServer[]> {
-    if (!fetchFromHost) {
-      return Promise.resolve(this.servers);  // Return the in-memory servers.
-    }
-    return this.digitalOcean.getDropletsByTag(SHADOWBOX_TAG).then((droplets) => {
-      this.servers = [];
-      return droplets.map((droplet) => {
-        return this.createDigitalOceanServer(this.digitalOcean, droplet);
-      });
-    });
-  }
-
-  // Creates a DigitaloceanServer object and adds it to the in-memory server list.
-  private createDigitalOceanServer(digitalOcean: DigitalOceanSession, dropletInfo: DropletInfo) {
-    const server = new DigitaloceanServer(digitalOcean, dropletInfo);
-    this.servers.push(server);
-    return server;
-  }
-}
-
-function sanitizeDigitaloceanToken(input: string): string {
-  const sanitizedInput = input.trim();
-  const pattern = /^[A-Za-z0-9_\/-]+$/;
-  if (!pattern.test(sanitizedInput)) {
-    throw new Error('Invalid DigitalOcean Token');
-  }
-  return sanitizedInput;
-}
-
-// cloudFunctions needs to define cloud::public_ip and cloud::add_tag.
-function getInstallScript(
-    accessToken: string, name: string, image?: string, watchtowerRefreshSeconds?: number,
-    metricsUrl?: string, sentryApiUrl?: string): string {
-  const sanitizedAccessToken = sanitizeDigitaloceanToken(accessToken);
-  // TODO: consider shell escaping these variables.
-  return '#!/bin/bash -eu\n' +
-      `export DO_ACCESS_TOKEN=${sanitizedAccessToken}\n` +
-      (image ? `export SB_IMAGE=${image}\n` : '') +
-      (watchtowerRefreshSeconds ?
-           `export WATCHTOWER_REFRESH_SECONDS=${watchtowerRefreshSeconds}\n` :
-           '') +
-      (sentryApiUrl ? `export SENTRY_API_URL="${sentryApiUrl}"\n` : '') +
-      (metricsUrl ? `export SB_METRICS_URL=${metricsUrl}\n` : '') +
-      `export SB_DEFAULT_SERVER_NAME="${name}"\n` + do_install_script.SCRIPT;
 }
